+++ include/lights/DirectionalLight.h
@@ -23,6 +23,7 @@ public:
 
     void sendToShader(Shader shader, std::string type) override;
 
+    DirectionalLight();
 };